+++ CHANGES.rst
@@ -3,6 +3,9 @@ Version 3.1.3
 
 Unreleased
 
+-   The session is marked as accessed for operations that only access the keys
+    but not the values, such as ``in`` and ``len``. :ghsa:`68rp-wp8r-4726`
+
 
 Version 3.1.2
 -------------